Moving utility classes from OpenXmlTemplate.Impl.cs to a separate file and other minor changes
